refactor: Move oidc config creation to oidc.rs
