+++ rust/operator-binary/src/security/oidc.rs
@@ -1,14 +1,20 @@
-use std::collections::{BTreeMap, HashSet};
+use std::collections::BTreeMap;
 
 use rand::{distributions::Alphanumeric, Rng};
 use snafu::{OptionExt, ResultExt, Snafu};
 use stackable_nifi_crd::NifiCluster;
 use stackable_operator::{
-    builder::meta::ObjectMetaBuilder, client::Client, k8s_openapi::api::core::v1::Secret,
-    kube::ResourceExt,
+    builder::meta::ObjectMetaBuilder,
+    client::Client,
+    commons::{
+        authentication::oidc::{AuthenticationProvider, DEFAULT_OIDC_WELLKNOWN_PATH},
+        tls_verification::{CaCert, TlsServerVerification, TlsVerification},
+    },
+    k8s_openapi::api::core::v1::Secret,
+    kube::{runtime::reflector::ObjectRef, ResourceExt},
 };
 
-use super::authentication::STACKABLE_ADMIN_USERNAME;
+use super::authentication::{NifiAuthenticationConfig, STACKABLE_ADMIN_USERNAME};
 
 const STACKABLE_OIDC_ADMIN_PASSWORD_KEY: &str = STACKABLE_ADMIN_USERNAME;
 
@@ -25,15 +31,24 @@ pub enum Error {
     },
 
     #[snafu(display(
-        "found existing secret [{}/{}], but key {} is missing",
-        name,
-        namespace,
-        STACKABLE_OIDC_ADMIN_PASSWORD_KEY
+        "found existing admin password secret {secret:?}, but the key {STACKABLE_OIDC_ADMIN_PASSWORD_KEY} is missing",
     ))]
-    OidcAdminPasswordKeyMissing { name: String, namespace: String },
+    OidcAdminPasswordKeyMissing { secret: ObjectRef<Secret> },
+
+    #[snafu(display("invalid OIDC endpoint URL"))]
+    InvalidOidcEndpoint {
+        source: stackable_operator::commons::authentication::oidc::Error,
+    },
+
+    #[snafu(display("failed to build the OIDC wellknown path"))]
+    InvalidOidcWellknownPath { source: url::ParseError },
+
+    #[snafu(display("Nifi doesn't support skipping the OIDC TLS verification"))]
+    NoOidcTlsVerificationNotSupported {},
 }
 
-/// Generate a secret containing the password for the admin user that can access the API. This admin user is the same as for SingleUser authentication.
+/// Generate a secret containing the password for the admin user that can access the API. This admin
+/// user is the same as for SingleUser authentication.
 pub(crate) async fn check_or_generate_oidc_admin_password(
     client: &Client,
     nifi: &NifiCluster,
@@ -46,17 +61,17 @@ pub(crate) async fn check_or_generate_oidc_admin_password(
         .context(OidcAdminPasswordSecretSnafu)?
     {
         Some(secret) => {
-            let keys = secret
+            let admin_password_present = secret
                 .data
-                .unwrap_or_default()
-                .into_keys()
-                .collect::<HashSet<_>>();
-            if keys.contains(STACKABLE_OIDC_ADMIN_PASSWORD_KEY) {
+                .iter()
+                .flat_map(|data| data.keys())
+                .any(|key| key == STACKABLE_OIDC_ADMIN_PASSWORD_KEY);
+
+            if admin_password_present {
                 Ok(false)
             } else {
                 OidcAdminPasswordKeyMissingSnafu {
-                    name: build_oidc_admin_password_secret_name(nifi),
-                    namespace,
+                    secret: ObjectRef::from_obj(&secret),
                 }
                 .fail()?
             }
@@ -91,4 +106,60 @@ pub(crate) async fn check_or_generate_oidc_admin_password(
 
 pub fn build_oidc_admin_password_secret_name(nifi: &NifiCluster) -> String {
     format!("{}-oidc-admin-password", nifi.name_any())
+}
+
+pub fn add_oidc_config(
+    auth_config: &NifiAuthenticationConfig,
+    properties: &mut BTreeMap<String, String>,
+) -> Result<(), Error> {
+    let NifiAuthenticationConfig::Oidc { provider, oidc, .. } = auth_config else {
+        return Ok(());
+    };
+
+    let endpoint_url = provider
+        .endpoint_url()
+        .context(InvalidOidcEndpointSnafu)?
+        .join(DEFAULT_OIDC_WELLKNOWN_PATH)
+        .context(InvalidOidcWellknownPathSnafu)?;
+    properties.insert(
+        "nifi.security.user.oidc.discovery.url".to_string(),
+        endpoint_url.to_string(),
+    );
+    let (oidc_client_id_env, oidc_client_secret_env) =
+        AuthenticationProvider::client_credentials_env_names(&oidc.client_credentials_secret_ref);
+    properties.insert(
+        "nifi.security.user.oidc.client.id".to_string(),
+        format!("${{env:{oidc_client_id_env}}}").to_string(),
+    );
+    properties.insert(
+        "nifi.security.user.oidc.client.secret".to_string(),
+        format!("${{env:{oidc_client_secret_env}}}").to_string(),
+    );
+    let scopes = provider.scopes.join(",");
+    properties.insert(
+        "nifi.security.user.oidc.additional.scopes".to_string(),
+        scopes.to_string(),
+    );
+    properties.insert(
+        "nifi.security.user.oidc.claim.identifying.user".to_string(),
+        provider.principal_claim.to_string(),
+    );
+
+    if let Some(tls) = &provider.tls.tls {
+        let truststore_strategy = match tls.verification {
+            TlsVerification::None {} => NoOidcTlsVerificationNotSupportedSnafu.fail()?,
+            TlsVerification::Server(TlsServerVerification {
+                ca_cert: CaCert::SecretClass(_),
+            }) => "NIFI", // The cert get's added to the stackable truststore
+            TlsVerification::Server(TlsServerVerification {
+                ca_cert: CaCert::WebPki {},
+            }) => "JDK", // The cert needs to be in the system truststore
+        };
+        properties.insert(
+            "nifi.security.user.oidc.truststore.strategy".to_owned(),
+            truststore_strategy.to_owned(),
+        );
+    }
+
+    Ok(())
 }
